Restructure files and add runtime type checking
