+++ shared/events/amplitude/__init__.py
@@ -1,0 +1,3 @@
+from shared.events.amplitude.publisher import AmplitudeEventPublisher
+
+__all__ = ["AmplitudeEventPublisher"]
